backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
celery[redis]
redis
boto3
pydantic
pydantic-settings
python-dotenv
httpx
aiosmtplib
jinja2
python-multipart
email-validator
google-generativeai
azure-identity
azure-mgmt-resource
google-cloud-resource-manager
google-auth
cryptography
